moved code in UserProfile view to its own UserLanding component
