+++ src/components/ContributionGraph.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function ContributionGraph() {
+  return (
+    <div>
+      Graph goes here :)
+    </div>
+  );
+}
