+++ src/components/TaskWindow.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Container } from 'reactstrap';
+
+export default function TaskWindow() {
+  return (
+    <Container>
+      Task Window:
+    </Container>
+  );
+}
